+++ prefect_lakefs/__init__.py
@@ -1,4 +1,4 @@
 from . import _version
-from .blocks import LakefsBlock  # noqa
+from prefect_lakefs.credentials import LakeFSCredentials  # noqa F401
 
 __version__ = _version.get_versions()["version"]