+++ prefect_lakefs/credentials.py
@@ -1,20 +1,39 @@
-from typing import Optional
+from contextlib import contextmanager
+from typing import Generator, Optional, Union
 
-from lakefs_client.api.branches_api import BranchesApi
-from lakefs_client.api.commits_api import CommitsApi
-from lakefs_client.api.objects_api import ObjectsApi
+from lakefs_client import ApiClient
+from lakefs_client.apis import BranchesApi, CommitsApi, ObjectsApi
 from lakefs_client.client import LakeFSClient
 from lakefs_client.configuration import Configuration
-from prefect.blocks.abstract import CredentialsBlock
-from pydantic import Field
-from pydantic import SecretStr
+from prefect.blocks.core import Block
+from prefect.utilities.collections import listrepr
+from pydantic import Field, SecretStr
+from typing_extensions import Literal
+
+LAKEFS_CLIENT_TYPES = {
+    "branches": BranchesApi,
+    "commits": CommitsApi,
+    "objects": ObjectsApi,
+}
 
 
-class LakeFsCredentials(CredentialsBlock):
-    repository: str = Field(
-        description="The LakeFS repository to use.",
-        title="The LakeFS repository.",
-    )
+class LakeFSCredentials(Block):
+
+    """Credentials block for generating configured LakeFS API clients.
+
+    Attributes:
+        access_key_id: s3 compatible access_key_id to initiate lakeFS client.
+        secret_access_key: s3 compatible secret_access_key to initiate lakeFS client.
+        endpoint_url: http/https url hosting the lakefs API.
+
+    Example:
+        Load stored LakeFS credentials:
+        ```python
+        from prefect_lakefs.credentials import LakeFSCredentials
+
+        lakefs_credentials = LakeFSCredentials.load("BLOCK_NAME")
+        ```
+    """
 
     access_key_id: Optional[str] = Field(
         default=None,
@@ -32,61 +51,72 @@ class LakeFsCredentials(CredentialsBlock):
         title="LakeFS URL",
     )
 
-    branch: Optional[str] = Field(
-        default=None,
-        description="The LakeFS branch to use.",
-        title="LakeFS branch",
-    )
-    commit_on_task: bool = Field(
-        default=False,
-        description="Whether to commit every completed task.",
-        title="Commit on completed Task",
-    )
-    commit_on_flow: bool = Field(
-        default=False,
-        description="Whether to commit every completed flow.",
-        title="Commit on completed Flow",
-    )
-    mirror_git_repo: bool = Field(
-        default=False,
-        description="Whether to include metadata from local git repo in LakeFS commit metadata",
-        title="Mirrow Git Repo",
-    )
-    strict_no_dirty: bool = Field(
-        default=False,
-        description="Whether to ensure all local git changes have been committed (no un-staged changes / dirty repo)",
-        title="Strict no dirty git repo",
-    )
+    _block_type_name = "LakeFS Credentials"
+    # TODO: add lakefs logo_url
+    _logo_url = "https://images.ctfassets.net/gm98wzqotmnx/08yCE6xpJMX9Kjl5VArDS/c2ede674c20f90b9b6edeab71feffac9/prefect-200x200.png?h=250"  # noqa
+    # TODO: add LakeFSCredentials doc slug
+    _documentation_url = "https://prefecthq.github.io"  # noqa
 
-    @property
-    def client(self) -> LakeFSClient:
+    @contextmanager
+    def get_client(
+        self,
+        client_type: Literal["branches", "commits", "objects"],
+    ) -> Generator[LakeFSClient, None, None]:
+        """Convenience method for retrieving a LakeFS API client for deployment resources.
+
+        Args:
+            client_type: The resource-specific type of LakeFS client to retrieve.
+
+        Yields:
+            An authenticated, resource-specific LakeFS API client.
+
+        Example:
+            ```python
+            from prefect_lakefs.credentials import LakeFSCredentials
+
+            with LakeFSCredentials.get_client("branches") as branches:
+                for branch in branches.list_branches(repository="example"):
+                    print(branch.id)
+            ```
         """
-        Gets a LakeFS client.
+
+        client_config = Configuration(
+            host=self.endpoint_url,
+            username=self.access_key_id,
+            password=self.secret_access_key.get_secret_value(),
+        )
+        with ApiClient(configuration=client_config) as generic_client:
+            try:
+                yield self.get_resource_specific_client(
+                    client_type,
+                    api_client=generic_client,
+                )
+            finally:
+                generic_client.rest_client.pool_manager.clear()
+
+    def get_resource_specific_client(
+        self,
+        client_type: str,
+        api_client: ApiClient,
+    ) -> Union[BranchesApi, CommitsApi, ObjectsApi]:
+        """
+        Utility function for configuring a generic LakeFS client.
+
+        Args:
+            client_type: The lakefs API client type for interacting with specific
+                LakeFS resources.
 
         Returns:
-            LakeFSClient.
+            LakeFSClient: An authenticated, resource-specific LakeFS Client.
+
+        Raises:
+            ValueError: If `client_type` is not a valid LakeFS API client type.
         """
-        username = self.access_key_id
-        password = self.secret_access_key
-        configuration = Configuration(
-            host=self.endpoint_url,
-            username=username,
-            password=password,
-        )
-        client: LakeFSClient = LakeFSClient(configuration)
-        return client
 
-    @property
-    def objects_api(self) -> ObjectsApi:
-        objects: ObjectsApi = self.client.objects
-        return objects
-
-    @property
-    def commits_api(self) -> CommitsApi:
-        commits: CommitsApi = self.client.commits
-        return commits
-
-    @property
-    def branches_api(self) -> BranchesApi:
-        branches: CommitsApi = self.client.branches
-        return branches
+        try:
+            return LAKEFS_CLIENT_TYPES[client_type](api_client=api_client)
+        except KeyError:
+            raise ValueError(
+                f"Invalid client type provided '{client_type}'."
+                f" Must be one of {listrepr(LAKEFS_CLIENT_TYPES.keys())}."
+            )